+++ src/main/java/networkexp/Event.java
@@ -4,9 +4,9 @@ package networkexp;
  * Created by Dandoh on 6/27/17.
  */
 public abstract class Event {
-    public int time;
+    public long time;
 
-    public Event(int time) {
+    public Event(long time) {
         this.time = time;
     }
